requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.7.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-property==0.2.2
asyncstdlib==3.13.0
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
blinker==1.9.0
Brotli==1.1.0
browser-cookie3==0.20.1
bs4==0.0.2
cachetools==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cmake==3.30.1
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
curl_cffi==0.7.4
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
deep-sort-realtime==1.3.2
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
Django==5.0.7
django-ipware==7.0.1
et-xmlfile==1.1.0
executing==2.1.0
fake-useragent==2.0.3
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
Flask==3.1.0
flask-babel==4.0.0
Flask-JWT-Extended==4.7.1
Flask-MySQLdb==2.0.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.33.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==5.9.1
gradio_client==1.5.2
grpcio==1.65.4
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jax==0.4.31
jaxlib==0.4.31
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
Js2Py==0.74
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.4.1
kiwisolver==1.4.5
langdetect==1.0.9
Levenshtein==0.26.1
libclang==18.1.1
lz4==4.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
mmh3==5.0.1
mpmath==1.3.0
multidict==6.1.0
mysqlclient==2.2.7
names==0.3.0
namex==0.0.8
narwhals==1.19.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.1
oauthlib==3.2.2
onnx==1.17.0
openai==1.58.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.13
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==3.20.3
psutil==6.1.1
pure_eval==0.2.3
py-arkose-generator==0.0.0.2
pyarrow==18.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybabel==0.0.0.dev0
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
PyExecJS==1.5.1
Pygments==2.18.0
pyjsparser==2.7.1
PyJWT==2.10.1
pymailtm==1.1.1
pyparsing==3.1.2
PyPasser==0.0.5
pypdfium2==4.30.1
pyperclip==1.9.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-doctr==0.10.0
python-dotenv==1.0.1
python-ipware==3.0.0
python-multipart==0.0.20
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
random-password-generator==2.2.0
random-username==1.0.2
RapidFuzz==3.11.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruff==0.8.4
safehttpx==0.1.6
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
selenium==4.27.1
semantic-version==2.10.0
shadowcopy==0.0.4
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.41.3
streamlit==1.41.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf2onnx==1.16.1
tf_keras==2.17.0
threadpoolctl==3.5.0
tls-client==1.0.1
toml==0.10.2
tomlkit==0.13.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
TwoCaptcha==0.0.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
undetected-chromedriver==3.5.5
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.26.3
watchdog==6.0.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
WMI==1.5.1
wrapt==1.14.1
wsproto==1.2.0
XlsxWriter==3.2.0
yarl==1.18.3
